project/requirements.txt
python-pptx
Pillow
pytesseract
google-genai
rich
simple-term-menu
python-dotenv